Do not check whether we can return to start at every step
Also includes some small refactoring and formatting

+++ aqmaps/src/main/java/uk/ac/ed/inf/aqmaps/drone/Route.java
@@ -127,7 +127,7 @@ public class Route {
     }
     
     // Battery too low, hence reading is unreliable
-    if (sensor.battery < 10) {
+    if (sensor.battery.doubleValue() < 10) {
       return "#000000";
     }
     
@@ -163,7 +163,7 @@ public class Route {
     }
    
     // Battery too low, hence reading is unreliable
-    if (sensor.battery < 10) {
+    if (sensor.battery.doubleValue() < 10) {
       return "cross";
     }
